requirements.txt
opencv-contrib-python
matplotlib
rshell
pyserial